fix yarn-prejob
