+++ dinky-gateway/src/main/java/org/dinky/gateway/yarn/YarnApplicationGateway.java
@@ -51,8 +51,6 @@ import cn.hutool.core.io.FileUtil;
  * @since 2021/10/29
  */
 public class YarnApplicationGateway extends YarnGateway {
-    private final String tmpConfDir = String.format(
-            "%s/tmp/%s", System.getProperty("user.dir"), UUID.randomUUID().toString());
 
     @Override
     public GatewayType getType() {
@@ -105,15 +103,4 @@ public class YarnApplicationGateway extends YarnGateway {
         return result;
     }
 
-    private File preparSqlFile() {
-        File tempSqlFile =
-                new File(String.format("%s/%s", tmpConfDir, configuration.get(CustomerConfigureOptions.EXEC_SQL_FILE)));
-        String sql = config == null ? "" : config.getSql();
-        FileUtil.writeString(Optional.ofNullable(sql).orElse(""), tempSqlFile.getAbsolutePath(), "UTF-8");
-        return tempSqlFile;
-    }
-
-    public boolean close() {
-        return FileUtil.del(tmpConfDir);
-    }
 }
