+++ dinky-gateway/src/main/java/org/dinky/gateway/yarn/YarnPerJobGateway.java
@@ -19,7 +19,9 @@
 
 package org.dinky.gateway.yarn;
 
+import cn.hutool.core.collection.CollUtil;
 import org.dinky.assertion.Asserts;
+import org.dinky.constant.CustomerConfigureOptions;
 import org.dinky.data.enums.GatewayType;
 import org.dinky.gateway.result.GatewayResult;
 import org.dinky.gateway.result.YarnResult;
@@ -66,6 +68,9 @@ public class YarnPerJobGateway extends YarnGateway {
 
         YarnResult result = YarnResult.build(getType());
         try (YarnClusterDescriptor yarnClusterDescriptor = createInitYarnClusterDescriptor()) {
+            yarnClusterDescriptor.addShipFiles(CollUtil.newArrayList(preparSqlFile()));
+            addConfigParas(
+                    CustomerConfigureOptions.EXEC_SQL_FILE, configuration.get(CustomerConfigureOptions.EXEC_SQL_FILE));
             ClusterClientProvider<ApplicationId> clusterClientProvider = yarnClusterDescriptor.deployJobCluster(
                     clusterSpecificationBuilder.createClusterSpecification(), jobGraph, true);
             ClusterClient<ApplicationId> clusterClient = clusterClientProvider.getClusterClient();
